tf-modules/aws: remove provider configuration
Remove provider configuration from the module to allow configuring the
provider configuration by the module user. This allows a root
configuration to have different provider configurations that can be used
to configure multiple instances of a module differently.

A user of these aws module should be able to use the same module to
create EKS clusters and ECR repositories in different regions.

Signed-off-by: Sunny <darkowlzz@protonmail.com>