tf-modules/gcp/gke: ignore node_config changes
When using workload identity, an empty node_config.oauth_scopes is set
at provision time. But after the provision, some default scopes are
added to the cluster. Reapplying the same config tries to replace the
existing scopes with empty scopes. This replacement requires the
existing cluster to be deleted and recreated. To avoid cluster
recreation between test runs, set a lifecycle rule to ignore any changes
in node_config.

Signed-off-by: Sunny <darkowlzz@protonmail.com>