Add testenv.Destroy for deleting resources
testenv.Destroy() can be used to delete resources from a terraform
configuration directory without going through the whole infrastructure
set up steps. This is to be used in CI to always run at the end to make
sure that everything has been destroyed after the test run.

In ideal case, this won't be needed, as the environment constructor,
testenv.New(), has signal handlers to gracefully delete the
infrastructure. But if the whole test process gets terminated, a new
process can just run testenv.Destroy() to perform the cleanup.

Signed-off-by: Sunny <darkowlzz@protonmail.com>