EKS: Disable cloudwatch logs & cluster encryption
Cloudwatch logs and cluster encryption are not needed for the current
test infrastructure needs.
It is also a workaround for a bug in aws-nuke which gets stuck because
the log group get's recreated by the EKS cluster.
Disabling cluster encryption helps reduce KMS usage and reduce cost. It
can be enabled if needed in the future.

Signed-off-by: Sunny <github@darkowlzz.space>