reaper: Add aws-nuke integration
To integrate aws-nuke in reaper, it is divided into multiple
components by separating the scan and delete operations. In order to
achieve this, some of the aws-nuke code is copied, modified and
extended.
The resource data scanned by aws-nuke is converted to the reaper's
native resource type to be listed in a coherent manner with the other
reaper providers and also support json output.
Applying the retention period/age filter to aws-nuke resources also
requires post processing of the resources after scanning and before
deleting. aws-nuke doesn't provide option for custom filters. The
scanned resources are passed through a custom filter, implemented as an
extension of the Nuke type, which understands the custom createdat
timestamp format and filters the resources accordingly.

Signed-off-by: Sunny <github@darkowlzz.space>