Add aws-github-actions terraform module
aws-github-actions terraform module helps create the IAM role and policy
with all the necessary permissions required for running the test
infrastructure, and also add the necessary secrets and variables to the
github repository. It expects GitHub OIDC to be registered as the
federated identity provider in the AWS account. More details can be
found in the README of the module.

Signed-off-by: Sunny <github@darkowlzz.space>